flask-backend/requirements.txt
Flask
flask_cors
numpy
pandas
requests
scikit-learn
Pillow
torch
torchvision
Gunicorn
